Moving to async for webhook. (Still fails with cyclic dependency)
